Need to keep annotation itself for class level as only that way we can define a proguard rule that keeps the name of the methods with that annotation.
Update README.md with new proguard rules that work.

Need to keep the annotation itself because as it is never read in code
it would be removed by proguard in the shrinking step (which is before
obfuscation) then during obfuscation it would not be there anymore to
work as the anchor point to keep the names.

Update gradle to latest.