Force configurable path segment replacement if not empty to start with / because when empty the whole path segment will be removed e.g. /test/<segment>/path when "segment" is "" becomes /test/path and not /test//path
Made changes to tests and samples for that.
Also verify configurable path segment is actually a string that is valid to be in an url path.

Fixed match index to handle case where empty configurable path segments at the end of the url did not match.

Changed simple glob to "..*" which enfoce that there must be one char making sure that placeholders do not create empty path elements.