Use the cartesian product to determine mergeable deeplink templates that can be represented by a single <intent-filter> entry ans thus save space in the manifest. Added tests for the different cases.
Generate the manifest file into the default location for resources in ksp for compartibility with caching and up-to-date checks. But create a separate task that will move the generated manifest somewhere else so that it does not get included in the `classes.jar` file.

Remove activityClassFqn from app module annotation, where it is not supported.